requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
opencv-python>=4.5.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
tqdm>=4.62.0
PyYAML>=6.0
tensorboard>=2.8.0
pillow>=8.3.0
albumentations>=1.0.0
face-alignment>=1.3.0
dlib>=19.22.0 